requirements.txt
colorzero==1.1
gpiozero==1.5.0
paho-mqtt==1.4.0
RPi.GPIO==0.6.5
